ifgtb-admin-prod-fix/requirements.txt
asgiref==3.5.2
backcall==0.2.0
black==22.3.0
CacheControl==0.12.11
cachetools==5.1.0
certifi==2020.12.5
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.2
cssselect2==0.6.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.13
django-classy-tags==3.0.1
django-cms==3.10.0
django-cors-headers==3.12.0
django-extensions==3.1.5
django-filer==2.2
django-formtools==2.3
django-js-asset==1.2.2
django-mptt==0.13.4
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.1.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==4.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==5.0.1
djangocms-video==3.0.0
djangorestframework==3.13.1
easy-thumbnails==2.8.1
fcm-django==1.0.11
filelock==3.7.0
firebase-admin==5.2.0
flake8==4.0.1
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud-core==2.3.0
google-cloud-firestore==2.5.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
html5lib==1.1
httplib2==0.20.4
identify==2.5.1
idna==3.3
ipython-genutils==0.2.0
lxml==4.8.0
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.3
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
pre-commit==2.19.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
proto-plus==1.20.5
protobuf==3.20.1
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pyfcm==1.5.4
pyflakes==2.4.0
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.6
pytz==2021.1
PyYAML==6.0
pyzmq==22.0.2
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.2
svglib==1.3.0
terminado==0.9.2
testpath==0.4.4
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
Unidecode==1.1.2
uritemplate==4.1.1
urllib3==1.26.3
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1